pyproject.toml
[tool.poetry]
name = "pippi"
version = "0.20.19"
description = ""
authors = ["sboutzen <sboutzen@gmail.com>"]
readme = "README.md"

packages = [
    { include = "pippi", from = "." },
    { include = "dags", from = "./dev" },
]

exclude = [
    "**/dev/logs/**/*",
    "**/opt/**/*"
]

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.3"
boto3 = "^1.26.109"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "5.11.0"
pytest = "^7.2.2"
apache-airflow = "2.5.1"
twine = "^4.0.2"
boto3-stubs = {extras = ["all"], version = "^1.26.109"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
